% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensPcorr.R
\name{DensPcorr}
\alias{DensPcorr}
\title{Dens-based approach for precision matrix estimation.}
\usage{
DensPcorr(
  data,
  select = FALSE,
  dens.level = 0.5,
  plateau.thresh = 0.01,
  Parcorr.est = NULL,
  lambda = NULL
)
}
\arguments{
\item{data}{Input data matrix of size n (observations) times p (variables).}

\item{select}{Whether to conduct the Dens-based selection. If FALSE, output
will only contain the estimated partial correlation list and precision
matrix list corresponding to the default tuning parameter series ranging
from 1e-8 to 0.6. If TRUE, the ouput will include the previous results and
the selected partial correlation matrix and percision matrix corresponding
to the specified density level. Default is FALSE.}

\item{dens.level}{Specify the density level in Dens-based tuning parameter selection
method (0<dens.level<1). This option is valid only when select=TRUE.}

\item{plateau.thresh}{The criterion to select the plateau.}

\item{Parcorr.est}{Previous output from DensPcorr function.}

\item{lambda}{The tuning parameters for estimating the precision matrix
ranging from 0 to 1.}
}
\value{
An R list containing the following terms: \item{selected.precision}{Selected Precision matrix corresponding to dens.level.} \item{selected.lambda}{Selected tuning parameter corresponding to dens.level.}
}
\description{
Dens-based approach for precision matrix estimation.
}
\details{
This function implements the statistical method proposed in Wang et al. (2016).
See Rpackage "DensParcorr".
}
