% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNKF.filter.R
\name{DNKF.filter}
\alias{DNKF.filter}
\title{Differential Network Knockoff Filter}
\usage{
DNKF.filter(AD, NC, alpha, screen_ratio = 0.5, B = 10, fdr = 0.1, offset = 0)
}
\arguments{
\item{AD}{AD sample data. n-p-q array or A list of length n, each of which is
p-by-q spatial-temporal data matrix. n denotes the number of samples, p
denotes the spatial dimension, q denotes the temporal dimension.}

\item{NC}{NC sample data. n-p-q array or A list of length n, each of which is
p-by-q spatial-temporal data matrix. n denotes the number of samples, p
denotes the spatial dimension, q denotes the temporal dimension.}

\item{alpha}{Vector. The elasticnet mixing parameter, with \eqn{0\le\alpha\le 1}.}

\item{screen_ratio}{Screen ratio. When screening, the number of screen is
\eqn{ratio \cdot 0.5 \cdot p \cdot (p-1)}, (default: 0.5).}

\item{B}{Times of generating knockoff variables (default: 10)}

\item{fdr}{target false discovery rate (default: 0.1).}

\item{offset}{either 0 or 1 (default: 0).}
}
\value{
A list containing two lists, one of which is zero-one network and the
other is the weight network
}
\description{
Differential Network Knockoff Filter
}
\examples{
# set.seed(2023)
# p<-50; q<-30; n1<-n2<-30
# AD <- vector(mode = "list", length = n1)
# NC <- vector(mode = "list", length = n2)
# for (i in seq(n1)) {
#   AD[[i]] <- matrix(rnorm(p*q), nrow = p, ncol = q, byrow = TRUE)
# }
# for (i in seq(n2)) {
#   NC[[i]] <- matrix(rnorm(p*q), nrow = p, ncol = q, byrow = TRUE)
# }
# result <- DNKF.filter(AD, NC, alpha = c(0.2, 0.3))


}
