% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRho.R
\name{getRho}
\alias{getRho}
\title{Calculate Pearson Correlation Coefficient of X and y}
\usage{
getRho(X, y, N_screen)
}
\arguments{
\item{X}{n-by-p matrix of predictors.}

\item{y}{response vector of length n.}

\item{N_screen}{the number of columns left.}
}
\value{
a list containing a location vector denoting the location of the
screened out columns and the Matrix after screening.
}
\description{
Calculate Pearson Correlation Coefficient of X and y
}
