% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubscripts.R
\name{getSubscripts}
\alias{getSubscripts}
\title{Get the location of the edges}
\usage{
getSubscripts(v)
}
\arguments{
\item{v}{vector denoting the location of the upper triangular matrix elements
after vectorization}
}
\value{
a list containing the location of each element in the matrix before
vectorization.
}
\description{
Get the location of the edges
}
