% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{mut_in_f}
\alias{mut_in_f}
\title{Fisher Transformation}
\usage{
mut_in_f(x)
}
\arguments{
\item{x}{Data to be Fisher transformed.}
}
\value{
Data after Fisher transformation
}
\description{
Fisher Transformation
}
