% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{Accumulation function}
\usage{
plus(n)
}
\arguments{
\item{n}{a positive integer.}
}
\value{
a positive integer is equal to the accumulation from 1 to n.
}
\description{
Accumulation function
}
