% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.glmnet_coefdiff_bin.R
\name{stat.glmnet_coefdiff_bin}
\alias{stat.glmnet_coefdiff_bin}
\title{Importance statistics based on a GLM with cross-validation}
\usage{
stat.glmnet_coefdiff_bin(X, X_k, Zv, y, family = "binomial", ...)
}
\arguments{
\item{X}{n-by-p matrix of original variables.}

\item{X_k}{n-by-p matrix of knockoff variables.}

\item{Zv}{a vector or matrix, containing the covariates.}

\item{y}{vector of length n, containing the response variables.}

\item{family}{response type.The default response family is 'binomial', for a
logistic regression model.}

\item{...}{Other arguments that can be passed to glmnet.}
}
\value{
A vector of statistics \eqn{W} of length p.
}
\description{
Importance statistics based on a GLM with cross-validation
}
